Revert "Feature/#119"
