+++ app/app/_components/appHeader.tsx
@@ -1,8 +1,8 @@
 'use client';
+import { ClassAttributes, ButtonHTMLAttributes } from 'react';
 import { AiOutlineCheck, AiOutlineDownload, AiOutlineEdit, AiOutlineUpload } from 'react-icons/ai';
-import { PiCards } from 'react-icons/pi';
-import { BiQuestionMark } from 'react-icons/bi';
-import { IoSettingsOutline } from 'react-icons/io5';
+import { PiCardsThin } from 'react-icons/pi';
+import { FiSettings } from 'react-icons/fi';
 
 export function AppHeader({
   openFile,
@@ -20,57 +20,40 @@ export function AppHeader({
   return (
     <header className="print-hidden w-full z-30 select-none flex-none ">
       <nav className="flex bg-main justify-between md:justify-start">
-        <NabButton
-          title="ファイルを開く"
-          onClick={openFile}
-          text="開く"
-          icon={<AiOutlineUpload />}
-        />
-        <NabButton
-          title={IsEditing ? '編集を完了する' : '編集する'}
-          onClick={() => setIsEditing(!IsEditing)}
-          text={IsEditing ? '完了' : '編集'}
-          icon={IsEditing ? <AiOutlineCheck /> : <AiOutlineEdit />}
-        />
-        <NabButton title="保存" onClick={save} text="保存" icon={<AiOutlineDownload />} />
+        <NabButton title="ファイルを開く" onClick={openFile}>
+          <span className="md:block hidden">開く</span>
+          <AiOutlineUpload className="md:hidden" />
+        </NabButton>
+        <NabButton onClick={() => setIsEditing(!IsEditing)}>
+          <span className="md:block hidden">{IsEditing ? '完了' : '編集'}</span>
+          <div className="md:hidden">{IsEditing ? <AiOutlineCheck /> : <AiOutlineEdit />}</div>
+        </NabButton>
+        {/* <NabButton title="英文、日本語訳の色を変更します">表示変更</NabButton> */}
+        <NabButton onClick={save}>
+          <span className="md:block hidden">保存</span>
+          <AiOutlineDownload className="md:hidden" />
+        </NabButton>
         <NabButton
           title="フラッシュカードをスタート"
           onClick={() => {
             setIsShowFlashCards(true);
           }}
-          text="フラッシュカード"
-          icon={<PiCards />}
-        />
-        <NabButton title="設定" text="設定" icon={<IoSettingsOutline />} />
-        <NabButton
-          title="ヘルプ"
-          onClick={() => window.open('./help', '_blank')}
-          text="ヘルプ"
-          icon={<BiQuestionMark />}
-        />
+        >
+          <span className="md:block hidden">フラッシュカード</span>
+          <PiCardsThin className="md:hidden" />
+        </NabButton>
+        <NabButton>
+          <span className="md:block hidden">設定</span>
+          <FiSettings className="md:hidden" />
+        </NabButton>
       </nav>
     </header>
   );
 }
-function NabButton({
-  title,
-  onClick,
-  text,
-  icon,
-}: {
-  title?: string;
-  onClick?: () => void;
-  text?: string;
-  icon?: JSX.Element;
-}) {
-  return (
-    <button
-      title={title}
-      onClick={onClick}
-      className="p-4 rounded hover:bg-main-hover flex items-center"
-    >
-      {icon}
-      <span className="md:block hidden px-2">{text}</span>
-    </button>
-  );
+function NabButton(
+  props: JSX.IntrinsicAttributes &
+    ClassAttributes<HTMLButtonElement> &
+    ButtonHTMLAttributes<HTMLButtonElement>
+) {
+  return <button {...props} className="p-4 rounded hover:bg-main-hover" />;
 }
