+++ app/app/_components/flashCards.module.css
@@ -1,0 +1,28 @@
+.flashCards {
+  animation: show 1s ease 0s normal;
+}
+
+.hidden {
+  animation: hidden 1s ease 0s normal;
+}
+.minimize {
+  animation: minimize 1s ease 0s normal;
+}
+
+@keyframes show {
+  0% {
+    top: 100vh;
+  }
+  100% {
+    top: 2.5em;
+  }
+}
+@keyframes hidden {
+  100% {
+    top: 100vh;
+  }
+}
+@keyframes minimize {
+  to {
+  }
+}
