--- electron/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "view-english",
-  "version": "1.0.0",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/chakkun1121/view-english.git"
-  },
-  "bugs": {
-    "url": "https://github.com/chakkun1121/view-english/issues"
-  },
-  "homepage": "https://github.com/chakkun1121/view-english#readme",
-  "scripts": {
-    "start": "electron .",
-    "build": "electron-builder -l -w --x64"
-  },
-  "author": "chakkun1121",
-  "devDependencies": {
-    "electron": "latest",
-    "electron-builder": "^24.6.4"
-  }
-}
