+++ next.config.js
@@ -6,19 +6,6 @@ const nextConfig = withPWA({
   /* config options here */
   basePath: process.env.NODE_ENV === 'development' ? '' : '/view-english',
   output: 'export',
-  runtimeCaching: [
-    // すべてのファイルをキャッシュし、オフライン対応させる
-    {
-      urlPattern: /^https?.*/,
-      handler: 'NetworkFirst',
-      options: {
-        cacheName: 'offlineCache',
-        expiration: {
-          maxEntries: 200,
-        },
-      },
-    },
-  ],
 });
 
 module.exports = nextConfig;
