+++ tsconfig.json
@@ -27,7 +27,7 @@
     ".next/types/**/*.ts",
     "**/*.ts",
     "**/*.tsx"
-, "electron/index.js"  ],
+  ],
   "exclude": [
     "node_modules"
   ]
